Cleaner solution to detecting stop criteria
Added the "Are We There Yet?" function, that checks each impulse wether the target time or distance has been reached, and emits a different message to server.js, stopping the rowing session. In this set-up, sessionState is also set to "stopped", which is communicated to the BLE and PM5 devices.